added functionality to login and signup
