+++ client/src/models/User.ts
@@ -2,7 +2,5 @@
 
 export interface User {
     username: string | null;
-    email: string | null;
     password: string | null;
-    savedSnippet: [];
    }
