+++ client/src/pages/ScanSnippet.tsx
@@ -28,7 +28,6 @@ const ScanSnippet = () => {
                 variables: {
                     input:{
                         text: codeText,
-                        author: "meh",
                     }
                 },
             })
