--- client/src/utils/api.ts
@@ -1,64 +1,0 @@
-import type { User } from '../models/User.js';
-import type { Snippet } from '../models/Snippet.js';
-
-
-// route to get logged in user's info (needs the token)
-export const getMe = (token: string) => {
- return fetch('/api/users/me', {
-   headers: {
-     'Content-Type': 'application/json',
-     authorization: `Bearer ${token}`,
-   },
- });
-};
-
-
-export const createUser = (userData: User) => {
- return fetch('/api/users', {
-   method: 'POST',
-   headers: {
-     'Content-Type': 'application/json',
-   },
-   body: JSON.stringify(userData),
- });
-};
-
-
-export const loginUser = (userData: User) => {
- return fetch('/api/users/login', {
-   method: 'POST',
-   headers: {
-     'Content-Type': 'application/json',
-   },
-   body: JSON.stringify(userData),
- });
-};
-
-
-// save snippet data for a logged in user
-export const saveSnippet = (snippetData: Snippet, token: string) => {
- return fetch('/api/users', {
-   method: 'PUT',
-   headers: {
-     'Content-Type': 'application/json',
-     authorization: `Bearer ${token}`,
-   },
-   body: JSON.stringify(snippetData),
- });
-};
-
-
-// remove saved snippet data for a logged in user
-export const deleteSnippet = (snippetId: string, token: string) => {
- return fetch(`/api/users/snippets/${snippetId}`, {
-   method: 'DELETE',
-   headers: {
-     authorization: `Bearer ${token}`,
-   },
- });
-};
-
-
-export const searchOpenAi = (query: string) => {
- return fetch(`/api/openai/${query}`);
-};
